Bormin

+++ app/Http/Controllers/BookUserController.php
@@ -8,26 +8,58 @@ use Illuminate\Support\Facades\Validator;
 
 use App\Models\User;
 use App\Models\Book;
+use App\Models\LibraryRuleSet;
 
 class BookUserController extends Controller
 {
+
+    public function __construct() {
+        date_default_timezone_set('Asia/Bangkok');
+    }
+
+    public function getAllUsersBooks() {
+        $users = User::all();
+        foreach ($users as $user) {
+            $time_now = time();
+            $cost_per_day = LibraryRuleSet::where('name', 'default')->first()->cost_per_day_late_return;
+            
+            foreach ($user->books as $book) {
+                // $book->pivot->return_time = time() - 2*24*3600; // for testing expired
+                $days_late = ceil(($time_now - $book->pivot->return_time) / (24*3600));
+                if ($days_late > 0) {
+                    $book->expired = true;
+                    $book->days_past_expired = $days_late;
+                    $book->cost = $cost_per_day * $book->days_past_expired;
+                } else {
+                    $book->expired = false;
+                    $book->cost = 0;
+                }
+                $book->borrow_date = date('d.m.y', $book->pivot->borrow_time);
+                $book->return_date = date('d.m.y', $book->pivot->return_time);
+            }
+        }
+        return $users;
+    }
 
     public function getUserBooks() {
         $books = Auth::user()->books;
         $books->load('category');
         $time_now = time();
-        $cost_per_day = 1;
+        $cost_per_day = LibraryRuleSet::where('name', 'default')->first()->cost_per_day_late_return;
         
         foreach ($books as $book) {
-            $book->return_time = time() + 2*24*3600; // for testing expired
-            if ($book->return_time > $time_now) {
+            // $book->pivot->return_time = time() - 2*24*3600; // for testing expired
+            $days_late = ceil(($time_now - $book->pivot->return_time) / (24*3600));
+            if ($days_late > 0) {
                 $book->expired = true;
-                $book->days_past_expired = ($book->return_time - $time_now) / (24*3600);
+                $book->days_past_expired = $days_late;
                 $book->cost = $cost_per_day * $book->days_past_expired;
             } else {
                  $book->expired = false;
                  $book->cost = 0;
             }
+            $book->borrow_date = date('d.m.y', $book->pivot->borrow_time);
+            $book->return_date = date('d.m.y', $book->pivot->return_time);
         }
         return $books;
     }
@@ -35,7 +67,6 @@ class BookUserController extends Controller
     public function borrow(Request $request) {
         $validator = Validator::make($request->all(), [
             'book_id' => 'required|integer',
-            'duration' => ''
         ]);
 
         if ($validator->fails()) {
@@ -44,7 +75,7 @@ class BookUserController extends Controller
             ], 400);
         }
 
-        $duration = $request->duration || 7; // days
+        $duration = LibraryRuleSet::where('name', 'default')->first()->duration_per_borrow;
         $borrow_time = time();
         $return_time = $borrow_time + ($duration*24*3600);
 
@@ -58,17 +89,27 @@ class BookUserController extends Controller
                 ], 406);
         }
 
-        if ($user->books()->get()->count() >= 3) {
+        $borrow_limit = LibraryRuleSet::where('name', 'default')->first()->num_of_books_per_user;
+        if ($user->books()->get()->count() >= $borrow_limit) {
             // exceed borrowed books at a time
             return response()->json([
                     'error' => 'User has exceeded number of borrowed books.'
                 ], 406);
         }
 
+        if ($book->stock <= 0) {
+            return response()->json([
+                    'error' => 'This book is out of stock.'
+                ], 400);
+        }
+
         $user->books()->attach($request->book_id, [
             'borrow_time' => $borrow_time,
             'return_time' => $return_time
         ]);
+
+        $book->stock = $book->stock - 1;
+        $book->save();
 
         return response()->json([
             'message' => 'Successfully borrowed ' . $book->title . '!'
@@ -89,7 +130,23 @@ class BookUserController extends Controller
 
         $user = User::findOrFail($request->user_id);
         $book = Book::findOrFail($request->book_id);
+
+        $found = false; // find if user has actually borrowed  this book
+
+        foreach($user->books as $user_book) {
+            if ($user_book->id == $book->id)
+                $found = true;
+        }
+
+        if (!$found) {
+            return response()->json([
+                'error' => 'User did not borrow this book.'
+            ], 400);
+        }
+
         $user->books()->detach($book);
+        $book->stock = $book->stock + 1;
+        $book->save();
         
         return response()->json([
             'message' => 'Successfully returned ' . $book->title . '!'
